mobilender/requirements.txt
Django==3.1
django-filter==2.4.0
djangorestframework==3.11.1
drf-yasg==1.20.0